+++ skoolkit/trace.py
@@ -66,6 +66,7 @@ class Tracer(PagingTracer):
                     accept_int = True
                 if accept_int:
                     accept_int = simulator.accept_interrupt(registers, memory, pc)
+                    tstates = registers[25]
 
             pc = registers[24]
             operations += 1